pratical_round/Project/requirement.txt
certifi==2019.6.16
chardet==3.0.4
Django==1.11.21
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.1.4
idna==2.8
PyJWT==1.7.1
pytz==2019.1
requests==2.22.0
rest-condition==1.0.3
urllib3==1.25.3
